credential-management/fedcm-context.https.html

<!DOCTYPE html>
<title>Federated Credential Management API context tests.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>

<body>

<script type="module">
import {request_options_with_mediation_required,
        request_options_with_context,
        fedcm_test} from './support/fedcm-helper.sub.js';

fedcm_test(async t => {
  const cred = navigator.credentials.get(request_options_with_mediation_required());
  const title = await window.test_driver.get_fedcm_dialog_title();
  assert_true(title.toLowerCase().includes('sign in'));
  await cred;
}, "FedCM call defaults to 'signin' context.");

fedcm_test(async t => {
  const cred = navigator.credentials.get(request_options_with_context("manifest.py", "signup"));
  const title = await window.test_driver.get_fedcm_dialog_title();
  assert_true(title.toLowerCase().includes('sign up'));
  await cred;
}, "FedCM with 'signup' context.");

fedcm_test(async t => {
  const cred = navigator.credentials.get(request_options_with_context("manifest.py", "use"));
  const title = await window.test_driver.get_fedcm_dialog_title();
  assert_true(title.toLowerCase().includes('use'));
  await cred;
}, "FedCM with 'use' context.");

fedcm_test(async t => {
  const cred = navigator.credentials.get(request_options_with_context("manifest.py", "continue"));
  const title = await window.test_driver.get_fedcm_dialog_title();
  assert_true(title.toLowerCase().includes('continue'));
  await cred;
}, "FedCM with 'continue' context.");
</script>
